src/Parcorpus/Parcorpus.Services/Parcorpus.Services.AnnotationService/WordAlignerServer/requirements.txt
setuptools-rust
pytest
simalign
nltk
grpcio
protobuf